+++ publishable/database/seeds/PermissionsTableSeeder.php
@@ -12,6 +12,7 @@ class PermissionsTableSeeder extends Seeder
     {
         $keys = [
             'browse_admin',
+            'browse_bread',
             'browse_database',
             'browse_media',
             'browse_compass',
