+++ resources/views/tools/bread/index.blade.php
@@ -1,12 +1,10 @@
 @extends('voyager::master')
 
-@section('page_title', __('voyager.generic.viewing').' '.__('voyager.generic.database'))
+@section('page_title', __('voyager.generic.viewing').' '.__('voyager.generic.bread'))
 
 @section('page_header')
     <h1 class="page-title">
-        <i class="voyager-data"></i> {{ __('voyager.generic.database') }}
-        <a href="{{ route('voyager.database.create') }}" class="btn btn-success"><i class="voyager-plus"></i>
-            {{ __('voyager.database.create_new_table') }}</a>
+        <i class="voyager-bread"></i> {{ __('voyager.generic.bread') }}
     </h1>
 @stop
 
@@ -21,7 +19,7 @@
                     <thead>
                         <tr>
                             <th>{{ __('voyager.database.table_name') }}</th>
-                            <th style="text-align:right">{{ __('voyager.database.table_actions') }}</th>
+                            <th style="text-align:right">{{ __('voyager.database.bread_crud_actions') }}</th>
                         </tr>
                     </thead>
 
@@ -34,23 +32,31 @@
                                    data-name="{{ $table->name }}" class="desctable">
                                    {{ $table->name }}
                                 </a>
+                                <i class="voyager-data"
+                                   style="font-size:25px; position:absolute; margin-left:10px; margin-top:-3px;"></i>
                             </p>
                         </td>
 
-                        <td class="actions">
-                            <a class="btn btn-danger btn-sm pull-right delete_table @if($table->dataTypeId) remove-bread-warning @endif"
-                               data-table="{{ $table->name }}" style="display:inline; cursor:pointer;">
-                               <i class="voyager-trash"></i> {{ __('voyager.generic.delete') }}
-                            </a>
-                            <a href="{{ route('voyager.database.edit', $table->name) }}"
-                               class="btn btn-sm btn-primary pull-right" style="display:inline; margin-right:10px;">
-                               <i class="voyager-edit"></i> {{ __('voyager.generic.edit') }}
-                            </a>
-                            <a href="{{ route('voyager.database.show', $table->name) }}"
-                               data-name="{{ $table->name }}"
-                               class="btn btn-sm btn-warning pull-right desctable" style="display:inline; margin-right:10px;">
-                               <i class="voyager-eye"></i> {{ __('voyager.generic.view') }}
-                            </a>
+                        <td class="actions text-right">
+                            @if($table->dataTypeId)
+                                <a href="{{ route('voyager.' . $table->slug . '.index') }}"
+                                   class="btn btn-warning btn-sm browse_bread" style="margin-right: 0;">
+                                    <i class="voyager-plus"></i> Browse
+                                </a>
+                                <a href="{{ route('voyager.bread.edit', $table->name) }}"
+                                   class="btn btn-primary btn-sm edit">
+                                    <i class="voyager-edit"></i> {{ __('voyager.generic.edit') }}
+                                </a>
+                                <a href="#delete-bread" data-id="{{ $table->dataTypeId }}" data-name="{{ $table->name }}"
+                                     class="btn btn-danger btn-sm delete">
+                                    <i class="voyager-trash"></i> {{ __('voyager.generic.delete') }}
+                                </a>
+                            @else
+                                <a href="{{ route('voyager.bread.create', ['name' => $table->name]) }}"
+                                   class="_btn btn-default btn-sm pull-right">
+                                    <i class="voyager-plus"></i> {{ __('voyager.database.add_bread') }}
+                                </a>
+                            @endif
                         </td>
                     </tr>
                 @endforeach
@@ -58,25 +64,22 @@
             </div>
         </div>
     </div>
-
-    <div class="modal modal-danger fade" tabindex="-1" id="delete_modal" role="dialog">
+    {{-- Delete BREAD Modal --}}
+    <div class="modal modal-danger fade" tabindex="-1" id="delete_builder_modal" role="dialog">
         <div class="modal-dialog">
             <div class="modal-content">
                 <div class="modal-header">
                     <button type="button" class="close" data-dismiss="modal" aria-label="{{ __('voyager.generic.close') }}"><span
                                 aria-hidden="true">&times;</span></button>
-                    <h4 class="modal-title"><i class="voyager-trash"></i> {!! __('voyager.database.delete_table_bread_quest', ['table' => '<span id="delete_table_name"></span>']) !!}</h4>
+                    <h4 class="modal-title"><i class="voyager-trash"></i>  {!! __('voyager.bread.delete_table_bread_quest', ['table' => '<span id="delete_builder_name"></span>']) !!}</h4>
                 </div>
                 <div class="modal-footer">
-                    <form action="{{ route('voyager.database.destroy', ['database' => '__database']) }}" id="delete_table_form" method="POST">
+                    <form action="{{ route('voyager.bread.delete', ['id' => null]) }}" id="delete_builder_form" method="POST">
                         {{ method_field('DELETE') }}
                         <input type="hidden" name="_token" value="{{ csrf_token() }}">
-                        <input type="submit" class="btn btn-danger pull-right" value="{{ __('voyager.database.delete_table_confirm') }}">
-                        <button type="button" class="btn btn-outline pull-right" style="margin-right:10px;"
-                                data-dismiss="modal">{{ __('voyager.generic.cancel') }}
-                        </button>
+                        <input type="submit" class="btn btn-danger" value="{{ __('voyager.database.delete_table_bread_conf') }}">
                     </form>
-
+                    <button type="button" class="btn btn-outline pull-right" data-dismiss="modal">{{ __('voyager.generic.cancel') }}</button>
                 </div>
             </div><!-- /.modal-content -->
         </div><!-- /.modal-dialog -->
@@ -141,6 +144,17 @@
 
         $(function () {
 
+            // Setup Delete BREAD
+            //
+            $('table .actions').on('click', '.delete', function (e) {
+                id = $(this).data('id');
+                name = $(this).data('name');
+
+                $('#delete_builder_name').text(name);
+                $('#delete_builder_form')[0].action += '/' + id;
+                $('#delete_builder_modal').modal('show');
+            });
+
             // Setup Show Table Info
             //
             $('.database-tables').on('click', '.desctable', function (e) {
@@ -161,19 +175,6 @@
                         $('#table_info').modal('show');
                     });
                 });
-            });
-
-            // Setup Delete Table
-            //
-            $('td.actions').on('click', '.delete_table', function (e) {
-                table = $(this).data('table');
-                if ($(this).hasClass('remove-bread-warning')) {
-                    toastr.warning("Please make sure to remove the BREAD on this table before deleting the table.");
-                } else {
-                    $('#delete_table_name').text(table);
-                    $('#delete_table_form')[0].action = $('#delete_table_form')[0].action.replace('__database', table);
-                    $('#delete_modal').modal('show');
-                }
             });
         });
     </script>
