+++ tests/RouteTest.php
@@ -56,7 +56,7 @@ class RouteTest extends TestCase
             route('voyager.menus.show', ['menu' => 1]),
             route('voyager.menus.edit', ['menu' => 1]),
             route('voyager.database.index'),
-            route('voyager.database.bread.edit', ['table' => 'categories']),
+            route('voyager.bread.edit', ['table' => 'categories']),
             route('voyager.database.edit', ['table' => 'categories']),
             route('voyager.database.create'),
         ];
